Removed from pthread.h implementation details
