+++ library/include/pthread.h
@@ -26,8 +26,6 @@
 #include <sys/types.h>
 #include <sys/time.h>
 #include <errno.h>
-#include <exec/types.h>
-#include <exec/semaphores.h>
 #include <sched.h>
 #include <features.h>
 
@@ -144,12 +142,7 @@ struct pthread_mutexattr {
 
 typedef struct pthread_mutexattr pthread_mutexattr_t;
 
-struct pthread_mutex {
-    APTR mutex;
-    int kind;
-    int incond;
-    struct Task *owner;
-};
+struct pthread_mutex;
 
 typedef struct pthread_mutex pthread_mutex_t;
 
@@ -173,11 +166,7 @@ struct pthread_condattr {
 
 typedef struct pthread_condattr pthread_condattr_t;
 
-struct pthread_cond {
-    int pad1;
-    struct SignalSemaphore semaphore;
-    struct MinList waiters;
-};
+struct pthread_cond;
 
 typedef struct pthread_cond pthread_cond_t;
 
@@ -195,12 +184,7 @@ struct pthread_barrierattr {
 
 typedef struct pthread_barrierattr pthread_barrierattr_t;
 
-struct pthread_barrier {
-    unsigned int curr_height;
-    unsigned int total_height;
-    pthread_cond_t breeched;
-    pthread_mutex_t lock;
-};
+struct pthread_barrier;
 
 typedef struct pthread_barrier pthread_barrier_t;
 
@@ -214,9 +198,7 @@ struct pthread_rwlockattr {
 
 typedef struct pthread_rwlockattr pthread_rwlockattr_t;
 
-struct pthread_rwlock {
-    struct SignalSemaphore semaphore;
-};
+struct pthread_rwlock;
 
 typedef struct pthread_rwlock pthread_rwlock_t;
